moved all the helper functions from lxd/storage_zfs.go to lxd/storage_zfs_utils.go Closes: #3471
